+++ shared/CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+### v1.1.0
+
+#### 新規モデル追加
+
+- [19. 試合イベント(Match-Event-Type)]
+- [20. フォーメーション(Formation)]
+- [21. 監督・コーチ(Manager)]
+- [22. 選手の出場履歴(Player-Appearance)]
+- [23. 監督・コーチの出場履歴(Manager-Appearance)]
+- [24. 選手の試合イベントログ(Player-Match-Event-Log)]
+- [25. 監督・コーチの試合イベントログ(Staff-Match-Event-Log)]
+- [26. 試合でのフォーメーション(Match-Team-Formation)]
+
 ### v1.0.10
 
 #### README.md, CHANGELOG.md 追加